+++ packages/calendar/src/components/calendar-types/date-range-calendar/util/__tests__/DayStateFactory.test.ts
@@ -10,12 +10,17 @@ describe("DayStateFactory", () => {
       expect(state!["2018-12"][52]).toBeDefined();
       expect(state!["2018-12"][52][29].highlights).toEqual([
         "selected",
+        "selectedStart",
         "range"
       ]);
       expect(state!["2018-12"][52][30].highlights).toEqual(["range"]);
       expect(state!["2018-12"][1][31].highlights).toEqual(["range"]);
       expect(state!["2019-01"][1][1].highlights).toEqual(["range"]);
-      expect(state!["2019-01"][1][2].highlights).toEqual(["selected", "range"]);
+      expect(state!["2019-01"][1][2].highlights).toEqual([
+        "selected",
+        "selectedEnd",
+        "range"
+      ]);
     });
   });
 });