+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 #### Next release
 
 * Add `show` option to `WeekNumber` config used in `CalenderMonth`
-* Add new highlight states to `DayStateHighlight`: `selectedStart` and `selectedEnd`. For compatibility reasons the selected days hove both the old `selected` AND `selectedStart` and/or `selectedEnd` 
+* Add new highlight states to `DayStateHighlight`: `selectedStart` and `selectedEnd`. For compatibility reasons the selected days hove both the old `selected` AND `selectedStart` and/or `selectedEnd`
+* Add `borderColor` (and `borderColorDisabled`) to `ButtonTheme`
 
 #### 0.0.34