+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 #### Next release
 
+* Coming soon... 
+
+#### 0.0.35
+
+* Upgrade `date-fns` to `2.0.1`.
 * Add `show` option to `WeekNumber` config used in `CalenderMonth`
 * Add new highlight states to `DayStateHighlight`: `selectedStart` and `selectedEnd`. For compatibility reasons the selected days hove both the old `selected` AND `selectedStart` and/or `selectedEnd`
 * Add `borderColor` (and `borderColorDisabled`) to `ButtonTheme`